refactor(config Parsing): adds Location.hpp
previously part of Website.hpp
